+++ Mailer/app/Models/Generators/NapunkNewsfilterGenerator.php
@@ -6,8 +6,8 @@ class NapunkNewsfilterGenerator extends NewsfilterGenerator
 {
     public function process(array $values): array
     {
-        $this->articleLocker->setLockText('Ezt a cikket csak a Napunk előfizetői olvashatják végig.');
-        $this->articleLocker->setupLockLink('Csatlakozz hozzánk', 'https://predplatne.dennikn.sk/napunk-start');
+        $this->n3ArticleLocker->setLockText('Ezt a cikket csak a Napunk előfizetői olvashatják végig.');
+        $this->n3ArticleLocker->setLockLink('Csatlakozz hozzánk', 'https://predplatne.dennikn.sk/napunk-start');
 
         return parent::process($values);
     }