+++ Mailer/extensions/mailer-module/src/Models/ContentGenerator/Engine/TwigEngine.php
@@ -5,6 +5,7 @@ namespace Remp\MailerModule\Models\ContentGenerator\Engine;
 use Twig\Environment;
 use Twig\Extra\Intl\IntlExtension;
 use Twig\Loader\ArrayLoader;
+use Twig\Markup;
 
 class TwigEngine implements IEngine
 {
@@ -17,5 +18,10 @@ class TwigEngine implements IEngine
         $twig->addExtension(new IntlExtension());
 
         return $twig->render('index.html', $params);
+    }
+
+    public function markSafe(string $content): \Stringable|string
+    {
+        return new Markup($content, 'UTF-8');
     }
 }